added tests

+++ src/main/i5/las2peer/services/appService/AppServiceHelper.java
@@ -8,16 +8,11 @@ import org.slf4j.LoggerFactory;
 import javax.json.*;
 import javax.json.stream.JsonGenerator;
 import javax.ws.rs.core.Response;
-import java.io.ByteArrayOutputStream;
-import java.io.InputStream;
-import java.io.StringReader;
-import java.io.UnsupportedEncodingException;
+import java.io.*;
 import java.net.URI;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.Date;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.stream.Collector;
 
 /**
@@ -37,41 +32,41 @@ public class AppServiceHelper {
 
     private void touchUser(AppService.User user) throws SQLException {
         ResultSet rs = dm.query("SELECT * FROM users WHERE oidc_id=?", user.oidc_id);
-        if (rs.next())
-            if (!rs.getString("username").equals(user.username)) {
+        if (rs.next()) {
+            if (!rs.getString("username").equals(user.username))
                 dm.update("UPDATE users SET username=? WHERE oidc_id=?", user.username, user.oidc_id);
-            }
-        else
+        } else {
             dm.update("INSERT INTO users VALUES (?,?)", user.oidc_id, user.username);
+        }
     }
 
     private String toSearchText(String text) {
-        return text.replaceAll("[^a-zA-Z0-9]", " ").toLowerCase();
+        return text.replaceAll("[\t\n ]+"," ").replaceAll("[^a-zA-Z0-9 ]", "").toLowerCase();
     }
     private String extractPlatforms(Map<String, Object> config) {
         StringBuilder sb = new StringBuilder();
-        for (String platform : ((Map<String, Object>) config.get("export")).keySet())
+        for (String platform : ((Map<String, Object>) config.getOrDefault("export", new HashMap<>())).keySet())
             sb.append(platform+";");
         return sb.toString();
     }
     public Response addApp(String description, Map<String, Object> config, AppService.User user) {
         try {
             touchUser(user);
-            dm.update("INSERT INTO apps VALUES (?,?,?,?,?,?)"
-                , null
-                , user.oidc_id
-                , description
-                , toSearchText(description)
-                , extractPlatforms(config)
-                , toJsonString(config)
-            );
-            ResultSet rs = dm.query("SELECT MAX(app) FROM apps");
+            ResultSet rs = dm.update("INSERT INTO apps VALUES (?,?,?,?,?,?)"
+               , null
+               , user.oidc_id
+               , description
+               , toSearchText(description)
+               , extractPlatforms(config)
+               , toJsonString(config)
+            ).generated;
             rs.next();
-            dm.update("INSERT INTO maintainers VALUES (?,?)", rs.getInt("app"), user.oidc_id);
+            int app = rs.getInt(1);
+            dm.update("INSERT INTO maintainers VALUES (?,?)", app, user.oidc_id);
 
-            return Response.created(URI.create("http://./"+rs.getInt("app"))).build();
+            return Response.created(URI.create("http://./"+app)).entity("{\"app\":"+app+"}").build();
         } catch (SQLException e) {
-            l.error(e.toString());
+            StringWriter sw = new StringWriter();e.printStackTrace(new PrintWriter(sw));l.error(sw.toString());
         }
         return Response.serverError().build();
     }
@@ -90,7 +85,7 @@ public class AppServiceHelper {
             );
             return Response.ok().build();
         } catch (SQLException e) {
-            l.error(e.toString());
+            StringWriter sw = new StringWriter();e.printStackTrace(new PrintWriter(sw));l.error(sw.toString());
         }
         return Response.serverError().build();
     }
@@ -98,12 +93,14 @@ public class AppServiceHelper {
         try {
             touchUser(user);
             ResultSet rs = dm.query("SELECT creator FROM apps WHERE app=?", app);
+            if (!rs.next())
+                return Response.status(404).build();
             if (!rs.getString("creator").equals(user.oidc_id))
                 return Response.status(403).build();
             dm.update("DELETE FROM apps WHERE app=?", app);
             return Response.ok().build();
         } catch (SQLException e) {
-            l.error(e.toString());
+            StringWriter sw = new StringWriter();e.printStackTrace(new PrintWriter(sw));l.error(sw.toString());
         }
         return Response.serverError().build();
     }
@@ -112,32 +109,50 @@ public class AppServiceHelper {
             ResultSet rs = dm.query("SELECT description,config,username FROM apps JOIN users ON creator=oidc_id WHERE app=?", app);
             if (!rs.next())
                 return Response.status(404).build();
+            ResultSet rs2 = dm.query("SELECT AVG(CAST(value AS DOUBLE)) FROM ratings WHERE app=?", app);
+            rs2.next();
             return Response.ok("{" +
-                    "\"creator\":" + rs.getString("username") +
-                    ",\"description\":" + rs.getString("description") +
+                    "\"creator\":\"" + rs.getString("username") + "\"" +
+                    ",\"description\":\"" + rs.getString("description") + "\"" +
                     ",\"config\":" + rs.getString("config") +
+                    ",\"rating\":" + rs2.getDouble(1) +
                 "}").build();
         } catch (SQLException e) {
-            l.error(e.toString());
+            StringWriter sw = new StringWriter();e.printStackTrace(new PrintWriter(sw));l.error(sw.toString());
+        }
+        return Response.serverError().build();
+    }
+
+    public Response addMaintainer(int app, String oidc_id, AppService.User user) {
+        try {
+            touchUser(user);
+            if (!dm.query("SELECT * FROM maintainers WHERE app=? AND maintainer=?", app, user.oidc_id).next())
+                return Response.status(403).build();
+            if (dm.query("SELECT * FROM maintainers WHERE app=? AND maintainer=?", app, oidc_id).next())
+                return Response.ok().build();
+            dm.update("INSERT INTO maintainers VALUES (?,?)", app, oidc_id);
+            return Response.ok().build();
+        } catch (SQLException e) {
+            StringWriter sw = new StringWriter();e.printStackTrace(new PrintWriter(sw));l.error(sw.toString());
         }
         return Response.serverError().build();
     }
 
     public Response searchApp(String query) {
-        StringBuilder sqlQuery = new StringBuilder("SELECT (app,description) FROM apps WHERE TRUE");
+        StringBuilder sqlQuery = new StringBuilder("SELECT app,description FROM apps WHERE TRUE");
         for (String k : query.split(" ")) {
             sqlQuery.append(" AND ");
             if (k.length() < 4)
-                sqlQuery.append("search_text LIKE "+k);
+                sqlQuery.append("search_text REGEXP '"+k+"'");
             else {
                 sqlQuery.append("search_text REGEXP '.^"); // .^ never matches
                 k = k.replaceAll("[^a-zA-Z0-9]", " ").toLowerCase();
                 // edit distance 1: /cat/ becomes /(c?.?at)|(ca?.?t)|(cat?.?)/
-                for (int i=1; i<k.length(); i++)
+                for (int i=1; i <= k.length(); i++)
                     sqlQuery.append("|("+k.substring(0,i)+"?.?"+k.substring(i)+")");
                 // letter swaps: /cat/ becomes /(.ct)|(c.a)/
-                for (int i=1; i<k.length()-1; i++)
-                    sqlQuery.append("|("+k.substring(0,i-1)+"."+k.charAt(i-1)+k.substring(i)+")");
+                for (int i=1; i < k.length(); i++)
+                    sqlQuery.append("|("+k.substring(0,i-1)+"."+k.charAt(i-1)+k.substring(i+1)+")");
                 sqlQuery.append("'");
             }
         }
@@ -148,14 +163,14 @@ public class AppServiceHelper {
             jg.writeStartArray();
             while (rs.next()) {
                 jg.writeStartObject()
-                    .write("id", rs.getInt("app"))
+                    .write("app", rs.getInt("app"))
                     .write("description", rs.getString("description"))
                 .writeEnd();
             }
             jg.writeEnd().close();
             return Response.ok(baos.toString("utf8")).build();
         } catch (SQLException | UnsupportedEncodingException e) {
-            l.error(e.toString());
+            StringWriter sw = new StringWriter();e.printStackTrace(new PrintWriter(sw));l.error(sw.toString());
         }
         return Response.serverError().build();
     }
@@ -170,7 +185,7 @@ public class AppServiceHelper {
             if (platform.equals("all"))
                 rs = dm.query("SELECT app FROM apps");
             else
-                rs = dm.query("SELECT app FROM apps WHERE platform LIKE ?", platform);
+                rs = dm.query("SELECT app FROM apps WHERE platform REGEXP ?", platform);
             ByteArrayOutputStream baos = new ByteArrayOutputStream();
             JsonGenerator jg = Json.createGenerator(baos);
             jg.writeStartArray();
@@ -179,7 +194,7 @@ public class AppServiceHelper {
             jg.writeEnd().close();
             return Response.ok(baos.toString("utf8")).build();
         } catch (SQLException | UnsupportedEncodingException e) {
-            l.error(e.toString());
+            StringWriter sw = new StringWriter();e.printStackTrace(new PrintWriter(sw));l.error(sw.toString());
         }
         return Response.serverError().build();
     }
@@ -191,18 +206,18 @@ public class AppServiceHelper {
             dm.update("INSERT INTO comments VALUES (?,?,?,?)", app, user.oidc_id, now, text);
             return Response.ok().build();
         } catch (SQLException e) {
-            l.error(e.toString());
+            StringWriter sw = new StringWriter();e.printStackTrace(new PrintWriter(sw));l.error(sw.toString());
         }
         return Response.serverError().build();
     }
     public Response deleteComment(int app, int timestamp, String text, AppService.User user) {
         try {
             touchUser(user);
-            if (0 == dm.update("DELETE FROM comments WHERE (app,creator,timestamp)=(?,?,?)", app, user.oidc_id, timestamp))
+            if (0 == dm.update("DELETE FROM comments WHERE (app,creator,timestamp)=(?,?,?)", app, user.oidc_id, timestamp).rows)
                 return Response.status(404).build();
             return Response.ok().build();
         } catch (SQLException e) {
-            l.error(e.toString());
+            StringWriter sw = new StringWriter();e.printStackTrace(new PrintWriter(sw));l.error(sw.toString());
         }
         return Response.serverError().build();
     }
@@ -215,12 +230,13 @@ public class AppServiceHelper {
             while (rs.next())
                 jg.writeStartObject()
                     .write("creator",rs.getString("username"))
-                    .write("timestamp",rs.getString("timestamp"))
+                    .write("timestamp",rs.getInt("timestamp"))
                     .write("text",rs.getString("text"))
                 .writeEnd();
             jg.writeEnd().close();
-        } catch (SQLException e) {
-            l.error(e.toString());
+            return Response.ok(baos.toString("utf8")).build();
+        } catch (SQLException | UnsupportedEncodingException e) {
+            StringWriter sw = new StringWriter();e.printStackTrace(new PrintWriter(sw));l.error(sw.toString());
         }
         return Response.serverError().build();
     }
@@ -236,7 +252,7 @@ public class AppServiceHelper {
                 dm.update("INSERT INTO ratings VALUES (?,?,?)", app, user.oidc_id, value);
             return Response.ok().build();
         } catch (SQLException e) {
-            l.error(e.toString());
+            StringWriter sw = new StringWriter();e.printStackTrace(new PrintWriter(sw));l.error(sw.toString());
         }
         return Response.serverError().build();
     }
@@ -249,16 +265,18 @@ public class AppServiceHelper {
             dm.update("INSERT INTO media VALUES (?,?,?,?)", app, name, type, blob);
             return Response.ok().build();
         } catch (SQLException e) {
-            l.error(e.toString());
+            StringWriter sw = new StringWriter();e.printStackTrace(new PrintWriter(sw));l.error(sw.toString());
         }
         return Response.serverError().build();
     }
     public Response getMedia(int app, String name) {
         try {
             ResultSet rs = dm.query("SELECT * FROM media WHERE (app,name)=(?,?)", app, name);
+            if (!rs.next())
+                return Response.status(404).build();
             return Response.ok().type(rs.getString("type")).entity(rs.getBinaryStream("blob")).build();
         } catch (SQLException e) {
-            l.error(e.toString());
+            StringWriter sw = new StringWriter();e.printStackTrace(new PrintWriter(sw));l.error(sw.toString());
         }
         return Response.serverError().build();
     }
